pyproject.toml
[project]
name = "ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.7.0",
    "bitsandbytes>=0.46.0",
    "fastapi>=0.115.12",
    "intel-extension-for-pytorch>=2.7.0",
    "llama-cpp-python>=0.3.9",
    "ruff>=0.11.12",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.52.4",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "mypy>=1.16.0",
    "ruff>=0.11.12",
]

[tool.ruff]
exclude = [
    ".git",
    ".git-rewrite",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pycache__",
    "__pypackages__",
]
line-length = 120
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
# https://docs.astral.sh/ruff/rules
select = [
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "D",
    "UP",
    "ANN",
    "ASYNC",
    "S",
    "BLE",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "ISC",
    "ICN",
    "LOG",
    "G",
    "INP",
    "PIE",
    "T20",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "TCH",
    "INT",
    "ARG",
    "PTH",
    "ERA",
    "PGH",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "TRY",
    "FLY",
    "PERF",
    "RUF",
]
ignore = [
    "D101",
    "D104",
    "D107",
    "D100",
    "COM812",
    "T201",
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"typing.TypedDict".msg = "Use typing_extensions.TypedDict instead."

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.lint.isort]
force-to-top = ["typing"]
lines-after-imports = 2

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["S101", "S104", "S106"]

[tool.ruff.format]
indent-style = "space"
line-ending = "lf"
skip-magic-trailing-comma = false
docstring-code-format = false
docstring-code-line-length = "dynamic"
